Trivial: removed summary from table (obsolete in html5)
The summary attribute is obsolete. Consider describing the structure of the table in a caption element or in a figure element containing the table; or, simplify the structure of the table so that no description is needed.

Looked like the project was geared towards html5.
Ignore if you are being html agnostic.