Remove default `a` and `a:visited` color normalizations.
Inclusion of these normalizations can result in unwanted or
unexpected consequences. This is because `a:visited` has a
specificity of 0,0,1,1. Therefore, the pseudo-class has to be
included in every author link-style with a lower specificity.

This is not expected behaviour when building up link styles from
the browser defaults. Very little is lost by removing the link
color normalization.