[hidden] attribute should have high precedence
the [hidden] attribute should take precedence over other `display` declarations, otherwise conflicting information could be presented to assistive technologies. It wouldn't make sense for a screen reader to ignore an element marked as "hidden" while it's clearly visible on the page.

If the element is to be shown, the [hidden] attribute should be removed, not overridden.