fix: Remove blue on button, select on iOS 15+ Safari
On iOS 15+ devices the button text color has changed to blue
(via color: -apple-system-blue;).

This is a change from iOS 14- devices which displayed black. While all
browsers display buttons slightly differently I felt the blue change was
significant enough to warrant a reset as developers might not be
expecting this.